fix(geometry): mesh geometry attached directly to IfcBuilding (#1910)
Terrain/DGM exports hang an IfcShellBasedSurfaceModel straight off the
IfcBuilding rather than off a dedicated element. `is_non_geometric_spatial`
blocked IfcBuilding, so `has_geometry_by_name` returned false, the building
never became a geometry job, and the file loaded with correct metadata and
hierarchy while rendering nothing at all.

The reported (0,0,0) RTC offset is a consequence, not the cause: with no job to
sample, RTC detection had nothing to look at and fell through to the
placement-bounds scan, which sees only the origin placements these files use.
Once the building is sampled, the raw-vertex probe already in
`sample_element_translation` reads its first vertex and re-bases correctly, so
RTC detection itself needed no change.

IfcBuilding joins IfcSpace/IfcSite/IfcSpatialZone in the exempt set — the same
fix IfcSpatialZone got in #1075. The gate only permits meshing, so a building
with no representation (the common case) still produces nothing.

The schema_helpers tests move to their own module file to keep schema_helpers.rs
under the 400-line rule, matching the stream_meta / stream_meta_tests pattern.
